Integration von Rollen

+++ Project/src/main/java/com/team3/project/service/RoleService.java
@@ -1,16 +1,30 @@
 package com.team3.project.service;
 
 import com.team3.project.Classes.Enumerations;
-import com.team3.project.Classes.User;
+import com.team3.project.DAO.DAOUser;
 import com.team3.project.Classes.Role;
 import com.team3.project.DAO.DAORole;
 import com.team3.project.DAOService.DAORoleService;
+import com.team3.project.DAOService.DAOUserService;
+import com.team3.project.Websocket.WebsocketObserver;
 import jakarta.annotation.Nonnull;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.LinkedList;
 import java.util.List;
 
+@Service
 public class RoleService {
+
+    @Autowired
+    WebsocketObserver observer;
+    @Autowired
+    private Enumerations enumerations;
+
+    public RoleService() {
+
+    }
 
     /**
      * gets all Roles of one User
@@ -83,7 +97,16 @@ public class RoleService {
      */
     public void createVisualRole(@Nonnull final String newRoleName, @Nonnull final Enumerations.Role role) throws Exception{
         if(newRoleName.isEmpty()) throw new Exception("Empty new RoleName");
-        DAORoleService.create(newRoleName, getAuthorizationFromRoleAsInt(role));
+        boolean responce = DAORoleService.create(newRoleName, getAuthorizationFromRoleAsInt(role));
+        if (responce) {
+            List<DAORole> drl = DAORoleService.getAll();
+            DAORole dr = drl.get(drl.size()-1);
+            try {
+                observer.sendToRoleGroup(2, new Role(dr.getId(),dr.getName()));
+            } catch (Exception e) {
+                System.out.println("Observer nicht initialisiert");
+            }
+        }
     }
 
     /* Author: Henry L. Freyschmidt
@@ -95,8 +118,17 @@ public class RoleService {
      * delete visual Role
      * @param id - role id
      */
-    public void deleteVisualRoleById(int id){
-        DAORoleService.deleteById(id);
+    public void deleteVisualRoleById(int id) throws Exception {
+        DAORole dr = DAORoleService.getByID(id);
+        if (dr == null) throw new Exception("Role not found");
+        boolean responce = DAORoleService.deleteById(id);
+        if (responce) {
+            try {
+                observer.sendToRoleGroup(2, new Role(dr.getId(),dr.getName()));
+            } catch (Exception e) {
+                System.out.println("Observer nicht initialisiert");
+            }
+        }
     }
 
     /**
@@ -107,24 +139,38 @@ public class RoleService {
      */
     public void changeVisualRoleName(int id, @Nonnull final String newRoleName) throws  Exception{
         if(newRoleName.isEmpty()) throw new Exception("Empty new Role Name");
-        DAORoleService.updateNameById(id, newRoleName);
+        boolean responce = DAORoleService.updateNameById(id, newRoleName);
+        if (responce) {
+            //DAORole dr = DAORoleService.getByID(id);
+            try {
+                observer.sendToRoleGroup(2, new Role(id,newRoleName));
+            } catch (Exception e) {
+                System.out.println("Observer nicht initialisiert");
+            }
+        }
     }
 
     /**
-     * change visual Roles Name
-     * @param role - real Role
-     * @param user -  user which recieves visual Role
+     * add visual Role to User
      * @param visualRole- visual Role to be attributed
      * @throws Exception Null or Empty or out of domain Params
      */
-    public void saveVisualRole(Role visualRole, User user, Enumerations.Role role) throws Exception{
+    public void addVisualRole(int uID, Role visualRole) throws Exception{
         if(visualRole == null) throw new Exception("Null visual Role");
         if(visualRole.getID() < 0) throw new Exception("Invalid visual RoleID");
-        if(role == null) throw new Exception("Null Role");
-        if(user == null) throw new Exception("Null user");
-        if(user.getID() < 0) throw new Exception("Invalid UserID");
-        //DAOUserService.addRole(newRole, user.getID());
-
+        DAOUser du = DAOUserService.getWithAuthorizationById(uID);
+        if (du == null) throw new Exception("User not found");
+        DAORole dr = DAORoleService.getByID(visualRole.getID());
+        if (du == null) throw new Exception("User not found");
+        if (du.getAuthorization().getAuthorization() != dr.getAuthorizations().get(0).getAuthorization()) throw new Exception("Mismatching Authorisation");
+        boolean responce = DAOUserService.updateAddRoleById(du.getId(), dr);
+        if (responce) {
+            try {
+                observer.sendToProfileByID(2, visualRole);
+            } catch (Exception e) {
+                System.out.println("Observer nicht initialisiert");
+            }
+        }
     }
 
     /* Author: Henry van Rooyen
@@ -136,4 +182,36 @@ public class RoleService {
         return Enumerations.getAuthorizationFromRoleAsInt(role);
     }
 
+    public Enumerations.Role getRoleByInt(int roleID) throws Exception {
+        if (roleID > 4 || roleID < 1) throw new Exception("Invalid RoleID");
+        return Enumerations.getRoleFromInt(roleID);
+    }
+
+    public void addRolesToUser(int uID, LinkedList<Role> roleList) {
+        for (Role role : roleList){
+            try {
+                addVisualRole(uID,role);
+            } catch (Exception ignored){}
+        }
+    }
+
+    public List<Role> getAllVisualRoles() {
+        List<Role> rl = new LinkedList<Role>();
+        for(DAORole dr : DAORoleService.getAll()){
+            Role toAdd = new Role(dr.getId(),dr.getName());
+            toAdd.setAuth(Enumerations.getRoleFromInt(dr.getAuthorizations().get(0).getAuthorization()));
+            rl.add(toAdd);
+        }
+        return rl;
+    }
+
+    public static List<Role> toRoleList(List<DAORole> drl){
+        List<Role> rl = new LinkedList<Role>();
+        for (DAORole dr : drl){
+            Role toAdd = new Role(dr.getId(),dr.getName());
+            toAdd.setAuth(Enumerations.getRoleFromInt(dr.getAuthorizations().get(0).getAuthorization()));
+            rl.add(toAdd);
+        }
+        return rl;
+    }
 }
